recipe-app-data/src/main/java/sabillon/springframework5/recipe/app/data/converters/CategoryToCategoryCommand.java
/**
 * The $lock.
 */
package sabillon.springframework5.recipe.app.data.converters;

import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.stereotype.Component;
import sabillon.springframework5.recipe.app.data.commands.CategoryCommand;
import sabillon.springframework5.recipe.app.data.models.Category;

import java.util.Objects;

/**
 * The type Category to category command.
 */
@Component
public class CategoryToCategoryCommand implements Converter<Category, CategoryCommand> {

    @Synchronized
    @Override
    public CategoryCommand convert(Category source) {
        return !Objects.isNull(source) ? CategoryCommand.builder()
                .id(source.getId())
                .description(source.getDescription())
                .build() : null;
    }
}
